requirements.txt
pytest
flake8
flake8-builtins
flake8-isort
pymongo
pyyaml